requirements.txt
streamlit>=1.30.0
spacy>=3.7.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
PyMuPDF>=1.23.0
